rename: 상세뷰 캐러셀 디렉토리 이동
