+++ src/components/carousel/detailCarousel/detailCarousel.css.ts
