+++ src/components/carousel/detailCarousel/numberTag/NumberTag.tsx
